+++ lib/app/router/app_router.dart
@@ -3,6 +3,7 @@ import 'package:go_router/go_router.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../../features/auth/presentation/pages/login_page.dart';
+import '../../features/profile/presentation/pages/profile_page.dart';
 import '../../features/tasks/presentation/pages/tasks_page.dart';
 import '../router/shell_scaffold.dart';
 import 'route_names.dart';
@@ -58,8 +59,8 @@ class AppRouter {
           GoRoute(
             path: '/profile',
             name: RouteNames.profile,
-            pageBuilder: (context, state) => NoTransitionPage(
-              child: _placeholder('Profile'),
+            pageBuilder: (context, state) => const NoTransitionPage(
+              child: ProfilePage(),
             ),
           ),
         ],